Refactor light clustering
Move from high end to shared, so all renderers use clustering.
